requirements.txt
# Core dependencies
streamlit>=1.32.0
python-dotenv>=1.0.0
requests>=2.31.0

# LangChain and vector store
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
langchain-chroma>=0.0.1
chromadb>=0.4.22
jsonschema>=4.21.1

# Embeddings and ML
sentence-transformers>=2.2.2
torch>=2.2.0
transformers>=4.37.2

# Document processing
python-docx>=1.0.1
PyPDF2>=3.0.0

# Development and utilities
pytest>=8.0.0
black>=24.2.0
flake8>=7.0.0
runpod==1.5.0
fpdf==1.7.2